+++ src/lib/components/atoms/TaskDeleteButton.svelte
@@ -1,0 +1,23 @@
+<script lang="ts">
+	import { invalidateAll } from '$app/navigation';
+	import { trpc } from '$lib/trpcClient';
+
+	export let id: number;
+
+	const deleteTask = async (id: number) => {
+		try {
+			await trpc.task.delete.mutate(id);
+		} catch (error) {
+			console.error(error);
+		}
+
+		invalidateAll();
+	};
+</script>
+
+<button
+	type="button"
+	on:click={async () => await deleteTask(id)}
+	data-testid="delete-button"
+	class="group-hover:visible invisible"><ion-icon name="trash-outline" /></button
+>
